Working on new tile system

+++ core/src/org/rexellentgames/dungeon/util/file/FileReader.java
@@ -17,7 +17,7 @@ public class FileReader {
 		return this.stream.readBoolean();
 	}
 
-	public short readInt16() throws IOException {
+	public byte readInt16() throws IOException {
 		return this.stream.readShort();
 	}
